pyproject.toml
[tool.black]
line-length = 79
experimental-string-processing = true

[tool.isort]
line_length = 79
multi_line_output = 3
force_single_line = true
known_standard_library = [
]
known_third_party = [
]
known_local_folder = [
    "src",
]
sections = ['FUTURE', 'STDLIB', 'THIRDPARTY', 'FIRSTPARTY', 'LOCALFOLDER']
profile = "black"

[tool.pylint]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
]


[project]
name = "sendbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.1",
    "pillow>=11.2.1",
    "reportlab>=4.4.1",
]
